security tasks

+++ src/Controller/MeasurementController.php
@@ -16,6 +16,7 @@ class MeasurementController extends AbstractController
     #[Route('/', name: 'app_measurement_index', methods: ['GET'])]
     public function index(WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         return $this->render('measurement/index.html.twig', [
             'weather' => $weatherRepository->findAll(),
         ]);
@@ -24,6 +25,7 @@ class MeasurementController extends AbstractController
     #[Route('/new', name: 'app_measurement_new', methods: ['GET', 'POST'])]
     public function new(Request $request, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         $weather = new Weather();
         $form = $this->createForm(WeatherType::class, $weather, array('validation_groups' => array('NewWeather')));
         $form->handleRequest($request);
@@ -43,6 +45,7 @@ class MeasurementController extends AbstractController
     #[Route('/{id}', name: 'app_measurement_show', methods: ['GET'])]
     public function show(Weather $weather): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         return $this->render('measurement/show.html.twig', [
             'weather' => $weather,
         ]);
@@ -51,6 +54,7 @@ class MeasurementController extends AbstractController
     #[Route('/{id}/edit', name: 'app_measurement_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Weather $weather, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         $form = $this->createForm(WeatherType::class, $weather, array('validation_groups' => array('EditWeather')));
         $form->handleRequest($request);
 
@@ -69,6 +73,7 @@ class MeasurementController extends AbstractController
     #[Route('/{id}', name: 'app_measurement_delete', methods: ['POST'])]
     public function delete(Request $request, Weather $weather, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         if ($this->isCsrfTokenValid('delete'.$weather->getId(), $request->request->get('_token'))) {
             $weatherRepository->remove($weather, true);
         }
